--- packages/cra-template-seed/template/src/auth/Components/Validator.js
@@ -1,12 +1,0 @@
-import { useHasAccess } from '../../hooks/useAuthority';
-
-const ComponentValidator = ({ allowedAuthorities, authorityKey, children }) => {
-  const hasAccess = useHasAccess(allowedAuthorities, authorityKey);
-  if (!hasAccess) {
-    return null;
-  }
-
-  return children;
-};
-
-export default ComponentValidator;
